QQmlCodeModel: populate files after loading them
Populate the QML file after loading it, to avoid crashes later on.

UpdateItemInSnapshot would write the DomItem of the qml file inside the
snapshot, and then populate the lazy parts of the qml file while
checking whether the file should be added to the valid snapshot too.

This leaves the snapshot in an invalid state where the main
thread responding to requests might access and use the
only-partially-populated snapshot and crash, for example in the case of
the document symbols that expect the Qml file to not be empty.

Therefore, make sure that the qml file is populated by the qml file
loading thread before writing it inside the snapshot.

Note that addFileWatches() populates the file, so the crashes only
happened when the CMake feature is disabled. Add a test to make sure
that we don't load unpopulated files when the CMake feature is disabled.

Task-number: QTBUG-119163
Change-Id: I6e73420f750e873898b7b426fe333edaf4d6a7c0
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>