Prefer write via QQmlProperty in setInitialProperty and fix crash
Rely on QQmlProperty to properly resolve everything. Ideally, we could
get rid of the fallback loop, but QQmlProperty does not support setting
nested gadget properties.

Note that iniital properties are resolved in the context of the object,
not of the file call Qt.createObject – this is relevent when one wants
to set attached properties.

Change-Id: I3868aae47a9e507c3ab1b848c82512e5ee04ff18
Fixes: QTBUG-138825
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>