a11y: Send missing QAccessible::ObjectShow event for Qt Quick Controls
Before sending the QAccessibleEvent QAccessible::ObjectShow event from
QQuickItemPrivate::setEffectiveVisibleRecur() we checked if
QQuickItemPrivate::isAccessible was set. For e.g. Button, isAccessible
wasn't set by default: This was because isAccessible was only set as a
response to that an Accessible attached property was assigned to the
item.

Therefore, when a Button was shown it could end up not sending the
QAccessible::ObjectShow event.

This problem was first found on Button, but after more investigation, we
have found that the same problem applies to:

* BusyIndicator
* Button
* ComboBox
* Dial
* Label (without a text)
* ProgressBar
* RangeSlider
* RoundButton
* Slider
* TextArea
* TextField

The fix is to explicitly enable accessibility by calling
QQuickItemPrivate::setAccessible() on *all* Qt Quick Controls, since
they are all supposed to be accessible.

Change-Id: If4fa050149a31bf2baffa208716341d21631c042
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>