deliverHoverEventRecursive(): don't crash if a child disappears
Hover event delivery could result in deletion of a child item.
If this happens, just stop and wait until the next delivery, to avoid
dereferencing a dangling pointer.

In QTBUG-139561 an item's layer.enabled is bound to MouseArea's
containsMouse.  When the mouse enters, a QQuickShaderEffectSource is
created, and destroyed again when it exits; but at that time, the
QQuickShaderEffectSource is in the QList that
deliverHoverEventRecursive() is iterating. We must not try to visit it
(or access its itemPrivate, call isVisible(), etc.)  after it's
destroyed; but tracking each item's lifetime in this high-frequency
function would be expensive. So we have to hope that aborting delivery
here will not result in chronic hover failure in some other case; but so
far it seems this crash is very rare. Usually delivery of a hover event
to one item does not result in destruction of a sibling item.

Done-with: Lu YaNing <luyaning@uniontech.com>
Fixes: QTBUG-139561
Pick-to: 6.10 6.9 6.8
Change-Id: I0f761474ee07c9d6dd669702aff6905023242ac3
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: YaNing Lu <luyaning@uniontech.com>