QQmlDomAstCreator: enable scriptexpressions for DomCreationOption::Default
This commit enables scriptexpressions for DomCreationOptions::Default.

QQmlDomAstCreator does not know how to handle certain script
expressions, notably JS lists due to the weird way
PatternElementList::accept0() is implemented.

QQmlDomAstCreatorWithQQmlJSScope uses custom list iteration for types
like PatternElementList to make QQmlDomAstCreator support them
correctly.

Move the custom list iteration functionality out of
QQmlDomAstCreatorWithQQmlJSScope so that QQmlDomAstCreator can also
benefit from it.

Rename QQmlDomAstCreator to QQmlDomAstCreatorBase. The Base version does
not do any custom iteration and can be used by
QQmlDomAstCreatorWithQQmlJSScope (otherwise you end up doing the custom
iteration twice when adding it to QQmlDomAstCreatorBase).

Create a new class called QQmlDomAstCreator that inherits from
QQmlDomAstCreatorBase and uses the custom list iteration on top of it.

Add a test in tst_qmldomitem to make sure that a JS list (that
contains parts which are parsed as PatternElementList) is correctly
loaded in the DOM.

Furthermore, add minimal support for WithStatements in QQmlDomAstCreator
to avoid weird test failures because of writeoutchecks in
tst_qmlformat_cli::settings that uses a file with a with-statement.

Also fix the qmldom tool to correctly output "null" values in JSON
format by simplifying DomBase::dump()'s handling of values, and add a
test for it. Not outputting "null" leads to invalid JSON generation and
a failure in tst_qtquickview_signallistener at build time on android in
the CI.

Change-Id: I7f9258921d8515eb6b3f9cf5dbb2539ffb980aa7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>