Dom: also lazy-load files that don't exist on disk
Fix FileToLoad::fromMemory to not set an empty file path on non-existent
files. Otherwise, the lazy-loading machanism will see that the factory
has an empty path and will think that the factory is invalid and won't
populate the file.

Add a test that files with nonexistent file paths get populated. This
case might happen while writing tests with snippets, where no on-disk
file exists.

Change-Id: I48ffd5f47a5f5f5111cfd257059093ce13ed5293
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>