QQmlPropertyCache expand Override Semantics. QML path
This path focuses on the actual expansion of Override Semantics
with regard to newly added virtual and override keywords.

Since Override Semantics is central to the way PropertyCache is
being populated, expanding it requires additional care.
To ensure smooth and reliable iteration, this patch isolates
data-flows: within this patch adding properties from MetaObject
stay "unaffected" by the expansion, as well as the paths adding
Methods and Signals from the IR.
This is done with the help of 3 things:
1. Separating logging logic from the Override logic
2. Extracting Override Semantics into it's own namespace
3. Adding a nob to control the set of rules

To facilitate and separate testing of the logging logic and
override semantics, the principle of dependency injection is
being used, allowing us to easily Fake handling of override
semantics, simplifying testing of PropertyCache::appendPropertyAttr

Logging logic (messages) is kept on the side of QQmlPropertyCache
mostly because this entity logically has more context to do proper
logging. Taking into account how the class name and name of the
property are handled, those are kept to be "requested on demand",
a.k.a. query them only in cases when the logging indeed needs to be
performed.

Multiple append* methods have been moved to the private section to
avoid invalidation of caches. See comment above QQmlPropertyCache.

As a workaround to enable unit-testing of some methods,
tst_qqmlpropertycache has been added as a friend class.

This patch updates tests/auto/quick/qquicklistview/data/negativeDisplayMargin.qml to remove an unintended override of ListView’s built-in count property. The test declared its own count property, which implicitly shadowed ListView.count. The newly introduced validation correctly flagged this conflict, so the test has been adjusted accordingly.

Change-Id: I809e58e421fa50f5ee7f21b31edbf926ad472736
Task-number: QTBUG-98320
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>