SearchField: Keep currentIndex at -1 until user explicitly selects
Until now, currentIndex was assigned to different values on different
actions, like to 0 on model changes, or to the index of the highlighted
index according to the arrow key actions. These caused unexpected behavior
so the following changes were made:

- Removed setting to 0 when model is set in the beginning.
- Reset currentIndex to -1 on any user edit(typing/deleting)
- Disabled arrow key interactions on currentIndex

Now currentIndex only changes whether the user clicks an item in the
popup or presses Enter on a highlighted item.

Pick-to: 6.10
Change-Id: Ib1a991c87a767fe436b4b025a26834871401ca4b
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>