QQuickTextEdit a11y: Allow modifying text selection via a11y API
Introduce new QAccessibleQuickTextEdit as a11y implementation
for QQuickTextEdit.

Subclass the existing QAccessibleQuickItem and override the
methods to modify text selection.
(For retrieving the existing selection, the logic in
the base class is fine, but the TextEdit properties selectionStart
and selectionEnd are read-only, so cannot be set directly there.)

Implement the QAccessibleTextInterface::addSelection logic
in the base class, so it can also be reused by the
QAccessibleQuickTextInput class that will be added
in an upcoming commit to implement similar logic
for QQuickTextInput and QQuickTextField.

Extend the existing unit test to also test text selection.

Task-number: QTBUG-139943
Pick-to: 6.10 6.9 6.8
Change-Id: Ibd6aa3e1bc9a24352c9bef4fcf4a452252d6689c
Reviewed-by: Timon Sassor <timon.sassor@governikus.de>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>