Initialize QQuickGradientStop position to zero
The documentation for GradientStop position says that its default
value is zero.  However, the member is not initialized, so its
value is actually random.

This can cause all sorts of problems (included corrupted rendering
if the value happens to fall into a NaN range).

[ChangeLog][QtQuick][GradientStop] The position property is now
always initialized to 0.0; previously it was left uninitialized.

Pick-to: 6.10 6.9 6.8 6.5
Task-number: QTBUG-140364
Change-Id: I5dbc83838fb4d2f1a46c34ae61e25bfe3853eb16
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>