QQuickStyle: Be smarter about detecting if a theme is dark
On Windows 11, the colorScheme is 'Unknown' when a contrast theme is
applied. This is because contrast themes are seperate to the classic
light/dark mode settings, but share the similarity that they both
determine palette colors. Some of our QQC2 themes provide their own
colors (like the Basic style, and FluentWinUI3 style). These styles
provide different colors depending on whether they think the system is
in dark mode or not. However, since the colorScheme can be 'Unknown',
and officially still have a dark-ish system palette, we need to be a
little more intelligent.

Also consider the system to be in "dark mode", in cases where the
colorScheme is 'Unknown' and the windowText is brighter than the window
palette color roles.

Task-number: QTBUG-140507
Pick-to: 6.10
Change-Id: I179ce77104f0d682f353aec10d0f7f28660668b7
Reviewed-by: Doris Verria <doris.verria@qt.io>