Update controls theme on QGuiApp::styleHints palette or theme change
Previously, we would only update the theme for a style when
QGuiApplication::styleHints()->colorSchemeChanged() would be emitted.
This was unreliable in cases where the user changes the system theme to
a different theme that uses different colors, but the same color scheme.
In that case, the theme wouldn't update, resulting in stale palette
colors.

With 8e7b97bba513d3ba5526a1827bbda3470dd7d02c now forwarding theming
related events to QGuiApp::styleHints, we can install an event filter on
it, and react to both application palette changes and theme changes.

Fixes: QTBUG-140507
Pick-to: 6.11 6.10 6.8
Change-Id: I14df2b7c82a843e7a40cc17f587fc0de3c37c372
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>