Native Style: rework style and focus window initialization
QStyle is destroyed with the app singleton, but not re-created when a
new app is re-created. We change this to lazily initialize the style via
application static in QQuickStyleItem::style().

Furthermore the focus window is not re-connected to the new application
singleton, either. We lazily construct it and ensure that is cleaned up
when the application singleton is destroyed.

Pick-to: 6.10
Fixes: QTBUG-140220
Change-Id: I97281aaa91794a8a4e1c1fb3aa524dd54cf19709
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>