tst_qmlls_modules::qmldirImports: increase timeout
It seems that the timeout for this test is too short on a debug build
with ASAN, and leads to unrelated test failures on my machine. I
couldn't see any weird behavior in the profiler, apart from
QQmlJSUtils::resourceFilesFromBuildFolder() that might take more time
because it iterates <build>/qtbase/qml to search for .qrc files (and
depending on how many QML modules were built this might take less or
more time).

Change-Id: I3c7b0b6483c89de0a06460f8924943b598e5c0a6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>