FontDialogContent: Comply with accessibility insights for Windows
Accessibility Insights for Windows is a Windows application that will
check the UI accessibility tree for any potential issues or
anti-patterns.

The tool would complain about the FontDialog not giving the text fields
for font family, style and size an accessibility name.

In addition, the sample text edit was exposed as accessible, even
though it's an non-interactive part of the UI. We can simply hide it
from screen readers.

Pick-to: 6.10 6.8
Change-Id: Ice96528c3bcf018fabd95d5aaddd644b41bdfca1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>