FileDialog: Comply with accessibility insights for Windows
Accessibility Insights for Windows is a Windows application that will
check the UI accessibility tree for any potential issues or
anti-patterns.

The tool would complain about the FileDialog having a button without a
name, in the FolderBreadcrumbBar.

The SideBar buttonDelegate would use a Button with an IconImage
contentItem, which the tool would complain about, since the Button
control then wouldn't have a name.

The FileDialogDelegate would also expose every label that it uses to
QAccessible. I'd assume users of screen readers would find it
overwhelming that every delegate has more than 4 different labels that
are all exposed to external assistive technologies. The most important
item to expose is the delegate itself, the sub-labels showing the file
size and last modified date can probably be ignored.

Pick-to: 6.10 6.8
Change-Id: Iad3fbf4a4178934214028214947506eb9d8c2651
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>