make tst_qmlls_modules:qmldirImport more robust
The current situation leads to a race in qmlls: the file is opened and
then the build folder is announced to qmlls. Depending on whether the
file opening thread is fast enough, it might start processing the file
before processing the notification about the build folders, which leads
to random test failures of tst_qmlls_modules:qmldirImport (once every
three run on average on my machine).

Therefore, announce the build folder before opening the file.

Change-Id: Ic67da41f2bcbf18be6ce46ee48a771937150f479
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>