a11y: Report + allow modifying QQuickScrollBar position
Introduce QAccessibleQuickScrollBar as the a11y
implementation for QQuickScrollBar and override
the base class implementations for the
QAccessibleValueInterface methods to report and
allow setting the position of the scroll bar.
(The base class implementation would always use
meaningless values of 0.0.)

"Normalize" the valid values for
QQuickScrollBar::position which range
from 0.0 to 1.0 - QQuickScrollBar::position (see [1])
to a range of 0.0 to 100.0 for the accessible
interface.

In QQuickScrollBarPrivate::setPosition, send a
QAccessibleValueChangeEvent when the value
changes, so assistive technology are notified
of the new value.
This e.g. ensures that the value shown in
Accerciser's "Interface Viewer", section "Value"
gets updated when scrolling happens.

[1] https://doc.qt.io/qt-6/qml-qtquick-controls-scrollbar.html#position-prop

Fixes: QTBUG-140673
Task-number: QTBUG-139833
Change-Id: I4333d91a70c6056424fe1e09cb075cf6ca7266e2
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>