QtQml: Fix assigning to QML-defined list properties
Returning an invalid QVariant from toVariant() is not helpful. We can't
produce a QQmlListReference in that case because there is no object,
but a QObjectList works just fine.

Writing to list properties via the metaobject is very much supported.
We can easily implement it using the same technique we use for reading.

Finally, to get the signal counts right, we also need to compare the
lists before assignment when taking the QQmlProperty code path.

Pick-to: 6.10 6.8
Fixes: QTBUG-140690
Change-Id: I09d6412f3f2ec86144810f47d0d620b63dfcd280
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>