Doc: Use unambiguous link targets where possible
Some links to sections within the same document were incorrectly going
to other documents due to the way that QDoc handles duplicate section
names. Additionally, documents containing duplicate section titles are
not guaranteed to be handled as authors may expect.

Links were changed to be more specific or removed in the case where the
target was immediately after the original link.

Change-Id: I78696519bd811dfe264561b799898a1d94b027dc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>