Skip synthesis of QContextMenuEvent if eventpoint has a grab
In QQuickWindowPrivate::maybeSynthesizeContextMenuEvent(), if the mouse
event's QEventPoint already has a grab (passive or exclusive), it may be
that a TapHandler with default gesturePolicy has onTapped: menu.popup()
or a MouseArea has onClicked: menu.popup() or something like that.
We can at least detect that a MouseArea has an exclusive grab, or a
TapHandler that handles the right mouse button has a passive grab,
although we cannot see what it will do in the future when the right
mouse button is released. We don't want to flash up a menu on press if
the user's code is going to show a menu on release, so don't send the
QContextMenuEvent, just in case.

Amends 24ac76941e899a7a92c5c06f70bc2b9f0344daf0 and
9ac0e70cdbf99e06761b6edb28937fded8791bdb

[ChangeLog][Controls] To fix QTBUG-134903, when the right mouse button
is pressed, we don't send a QContextMenuEvent if we can detect that a
right-mouse click might already be handled by a TapHandler or a
MouseArea to open a Menu. But we recommend that you stop opening menus
that way: either rely on the built-in menus provided by TextArea and
TextField, or use the ContextMenu attached property to replace or add
menus to controls. In Qt 7 we plan to send the QContextMenuEvent without
checking for handlers first.

Fixes: QTBUG-134903
Task-number: QTBUG-132384
Change-Id: I1ef5387fbddeed4e6fc81d538683248a05d6e03f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>