Run tst_qquickfiledialogimpl also with the Imagine style
The intention was always to run it with every single built-in style.
Previously the Imagine style was omitted because of an issue that needed
more investigation. It seems like that has since been resolved, and that
it's safe to include it in the list of styles to run
tst_qquickfiledialogimpl with.

Change-Id: I7634a938c550392f8f5c18c141404c80a7ab99b2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>