mac style: update Switch
On macOS Tahoe 26, the native AppKit Switch has changed apparance
to become wider, with a liquid glass effect on the handle.

This patch will therefore update the Controls Switch to do the
same. That is, change the Switch to be equally wider if the app
is running with liquid glass, and change the appearance of the
handle to look a bit closer to the native handle.

Note that the Switch in Controls has always been drawn
'manually' with QML, so this change is not really fixing a
regression, but is more of a style update.

Task-number: QTBUG-138942
Change-Id: I1c7c9beb35845dac29c0fc67bd0813fffa313116
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>