QML Generator: Fix transform reference and improve layer handling
The recent mask commit mistakenly removed the generation of (layer)
transform references; this commit adds that back.

That commit also moved the generation of ids to the visitors. Hence we
can simplify the transform reference handling by just relating to
those ids, and drop the special layer number handling.

As a driveby, improve flexibility by replacing the hardcoded
"LayerItem" with a settable customItemType input field.

Change-Id: I694a7f5b274e8e114ea9a59f5331954a3e94b290
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>