qqmldiffer: adapt the differ from QtC
QtC has a well-tested diff calculator which works in the core of
DiffEditor plugin. Adapt it in qtdeclarative to be
used as qmlls utility by modernizing a little and removing the
unnecessary api that won't be needed. The first use case will be
shifting highlights based on the calculated differences. Also,
regexp-based highlighting would perform only on the diffed context
rather than the entire source code.

Task-number: QTBUG-140645
Change-Id: I5e88e4c6d77c1432b25eb226f9f5444928b9b156
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>