ScrollBar::test_flashing: Use non-interactive scrollbar to stabilize test
In rare cases the mouse was lingering over the exact position of the
scroll bar as a result of other tests moving the mouse, which would
make the scrollbar never leave the active state, and hence keep
reporting a non-0 opacity.

We're programmatically activating the scroll bar via increment(),
so we don't need the scroll bar to be interactive to do the test.

Pick-to: 6.10 6.8 6.5
Change-Id: Id75b22d51277c3515fd4f448360bd576b45d69d8
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>