Expose window devicePixelRatio as a property
On some platforms only the window knows the correct devicePixelRatio
which is different than the one on QScreen.
On QWindow changes to devicePixelRatio are exposed through an event
so we make forward the event to a signal here.
We hook it to QQuickWindow's effectiveDevicePixelRatio because
it takes into account QQuickRenderControl::renderWindowFor() and
QQuickRenderTarget::devicePixelRatio()
Effective device pixel ratio is returned as
it is the one that matters for rendering or calculations for sizes.
QQuickWidget forwards the events to the QQuickWindow already.

Fixes: QTBUG-141448
Change-Id: I2908b52dbcf005ec25ba282c92906eb6172c9d19
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>