QML generator: Avoid creating empty animations
When generating qml for an animated transform, we check for each
property if the value actually changes, and if not, we generate just
constant property assignment instead of an item to be animated. If it
turns out that the whole transform is constant, we would still
generate the animation block, but it would contain no actual value
animations. This is both redundant, and would cause a warning message
if the whole generated scene was paused ("setPaused() cannot be used
when animation isn't running").

Fix by only generating the animation block if any of the transform
property values is non-constant.

Change-Id: Iec9c7b580ab6245c46d6de8693f8c83a170d5049
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>