Ensure that the right grid at least have one column
If the initial window width was so large that the inital layout was
determined to be the "largeLayout" (wider than 450 pixels), then the
binding

    columns: Math.min(Math.round(width / 130), 6)

initially resulted in 0 columns because the width was initially
evaluated to be 0, which caused the grids implicitWidth to be 0 (because
it couldn't fit any items when columns was 0). Hence, it was
"deadlocked": it didn't add items because columns was 0, and it
didn't increase columns because it's width was 0.

An item in a layout with implicitWidth == 0 will always have lower
priority to grow than items with a larger implicitWidth - hence the grid
was stuck at width == 0

Pick-to: 6.10 6.8
Fixes: QTBUG-140533
Change-Id: I558463ac33db36fa7eb8df4c70adc94f90c69e8f
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>