Add tst_HoverHandler::twoHandlersTwoTouches
Clarify that the touch-hover behavior doesn't depend on touch->mouse
synthesis. We started faking hover on touchscreens because MouseArea can
handle synth-mouse events coming from touchscreens, so containsMouse was
reacting to touchpoints: then QTBUG-40856 came along, and the fix for it
cemented some rules in place. Then HoverHandler was added, and it was
deemed reasonable to get similar behavior, even though it handles touch
events directly: if your fingertip is touching, then the rest of your
finger is hovering (just as the mouse cursor is hovering at the same
time that the mouse button is pressed). In this test case, HoverHandler
directly handles touch events, and there is no touchmouse. So far, we
still have Qt 5 behavior: HoverHandler is a single-point handler, and if
you try to press two of them with two fingers, the first one "wins".
But this might need to change later, to support multi-finger hover.

Task-number: QTBUG-40856
Change-Id: If46d766d80534d729d254ca12ed1069c1d94e7a4
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>