Android: skip test cases failing to acquire deadlock protector
It crashes when failing to acquire UI thread deadlock protector
that's acquired by input context. The reason this happens is
that some QML tests have a TextEdit or TextInput based elements
that ends up doing some operation that blocks the UI thread
even after the test case is finished and then when a next
test case tries to create a window surface it finds the UI
thread blocked. Skip these for now, otherwise a solution
needs to be found under QTBUG-139400.

This includes:
* tst_qquickwidget::focusOnClickInProxyWidget
* tst_QQuickContextMenu::textControlsMenuKey()
* tst_QQuickContextMenu::mouseAreaUnderTextArea()

Fixes: QTBUG-141161
Task-number: QTBUG-139400
Pick-to: 6.10 6.8 6.5
Change-Id: I61eed34d79aaa88c46937610dec962dcecf5c716
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>
Reviewed-by: Soheil Armin <soheil.armin@qt.io>