qmllint: prefer enums over chained attached type
It seems that FlexboxLayout.Row is actually ambiguous: it could refer to
the "Row" enum value, and to the attached type of "Row" which is
attached on the attached type of "FlexboxLayout".

Fix qmllint to follow the qml engine's interpretation of the thing, and
resolve it as an enum instead of a chained attached type in
QQmlJSTypeResolver.

Pick-to: 6.10 6.8
Fixes: QTBUG-141194
Change-Id: I24e23f5fc92b0d007ff1d628b6286f85a60f10d8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>