QQuickOverlay: translate touch events that circumvent the deliveryagent
The touch event's scene position is normally set in the delivery agent.
However, since the overlay installs an event filter on the window
itself, it will circumvent the delivery agent and thus never set the
scene position properly.

This patch ensures that the scene position gets set when we're entering
overlayEvent from the event filter, for touch release events.

Task-number: QTBUG-132914
Pick-to: 6.10 6.8
Change-Id: Id59ed9e5252ba594ce0e40039cb3a783032caef1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>