Support masked items with transforms
When we get the bounds for our masked item, we need the bounds of the
actual shape it contains. This means we want to apply all styles except
the transform, since the transform will be applied to the ShaderEffect
item later on. All the other styles might affect the content bounds.

We could perhaps solve this by adding a contentBounds() to QSvgNode,
which returns the bounds untransformed, but for now we just apply
the inverted transform to the rect afterwards to cancel the effect.
This of course means that if you set a degenerate transform on the
item, it will not work correctly, but that implies that you are
scaling one of the axes by 0, so this is theoretical corner case.

Fixes: QTBUG-141168
Change-Id: I3221e599cf4bfe48733173ecb13e28058066b19c
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>