Binding: Fix "value changed" detection and restore logic
We need to keep in mind that the value might have changed compared to
what we initially had when writing to a property.

In a future dev only change, we might rather make Binding a
(conditional) interceptor. That would avoid the read and comparison, as
we could track invalidations more directly. However, for picking back to
stable branches, we want a more minimal change in any case.

This requires us to do a similar conversion as QQmlProperty::write does.
Moreover, we need to take additional care for QQmlListProperty: There,
we would so far store a QQmlListReference, referencing whatever the
Binding wrote to the property. Instead, we need to create a snapshot of
the list. Moreover, we can't just store a QList of QObjects, as the
objects might get garbage collected. Instead, put them into a
QQmlListWrapper (in "owning" mode).

Amends 6a8478829747289cdcce2a6e9628b31cfd865f15

Pick-to: 6.10 6.8
Fixes: QTBUG-140858
Change-Id: I5d0ec4ee5f1a20a6ee4b5f1696f561b2638bd07e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>