Stabilize tst_HoverHandler::movingItemWithHoverHandler
We know QCursor::setPos() doesn't work on Wayland; but let's detect
when it doesn't work, at runtime, and skip the rest of the test.

The reason for the change 1c7213dbc00fa595aeaa98ad5631bac56782599e was
that if there is no mouse, or if it never moves,
QGuiApplicationPrivate::lastCursorPosition might not have changed from
its initial offscreen position. So that's why this test does not send
any explicit mouse moves. But the window should receive an Enter event,
and QGuiAppPrivate::processEnterEvent() should set lastCursorPosition.
If that also didn't happen after showing the window with the cursor
above it, it's not Qt Quick's problem. flushFrameSynchronousEvents() may
not be able to hover and un-hover the moving item if it doesn't know
where the cursor is. If you have no mouse, the item can't be hovered
anyway (except that we might fake it with the touchmouse, but that is
not being tested here).

Fixes: QTBUG-76312
Fixes: QTBUG-81884
Fixes: QTBUG-103065
Fixes: QTBUG-111294
Change-Id: I76512ccf717385f32fe123b7c6898bff4eb17385
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>